refactor: rename database references
