+++ docker-compose.yaml
@@ -6,7 +6,7 @@ services:
     ports:
       - 5432:5432
     volumes:
-      - ./docker/postgres-data:/var/lib/postgresql/data
+      - ./docker/postgres:/var/lib/postgresql/data
     environment:
       POSTGRES_USER: ${DB_USER}
       POSTGRES_PASSWORD: ${DB_PASS}
