+++ projects/backend/src/core/session/repository/index.ts
@@ -1,3 +1,3 @@
 export type { SessionRepository } from './session.repository'
-export { SessionMysqlRepository } from './session.mysql.repository'
+export { SessionPostgresRepository } from './session.postgres.repository'
 export { SessionInMemoryRepository } from './session.in-memory.repository'
