+++ projects/backend/src/core/user/repository/index.ts
@@ -1,3 +1,3 @@
 export type { UserRepository } from './user.repository'
-export { UserMysqlRepository } from './user.mysql.repository'
+export { UserPostgresRepository } from './user.postgres.repository'
 export { UserInMemoryRepository } from './user.in-memory.repository'
